chore: rebuild gradle wrapper from sources

+++ src/main/kotlin/de/babsek/demo/axontesting/domain/events/MoneyTransferArrivedEvent.kt
@@ -1,7 +1,9 @@
 package de.babsek.demo.axontesting.domain.events
 
+import java.math.BigDecimal
+
 data class MoneyTransferArrivedEvent(
     val bankAccountId: String,
-    val amount: Double,
-    val reason: String
+    val amount: BigDecimal,
+    val reason: String,
 )
