+++ src/main/kotlin/de/babsek/demo/axontesting/domain/events/MoneyTransferFailedEvent.kt
@@ -1,8 +1,10 @@
 package de.babsek.demo.axontesting.domain.events
 
+import java.math.BigDecimal
+
 data class MoneyTransferFailedEvent(
     val bankAccountId: String,
     val targetBankAccountId: String,
-    val amount: Double,
-    val errorMessage: String?
+    val amount: BigDecimal,
+    val errorMessage: String?,
 )
