+++ src/main/kotlin/de/babsek/demo/axontesting/domain/events/MoneyTransferRequestedEvent.kt
@@ -1,8 +1,10 @@
 package de.babsek.demo.axontesting.domain.events
 
+import java.math.BigDecimal
+
 data class MoneyTransferRequestedEvent(
     val originBankAccountId: String,
     val targetBankAccountId: String,
-    val amount: Double,
-    val reason: String
+    val amount: BigDecimal,
+    val reason: String,
 )
