+++ src/main/kotlin/de/babsek/demo/axontesting/domain/value/TransactionDetails.kt
@@ -1,10 +1,11 @@
 package de.babsek.demo.axontesting.domain.value
 
+import java.math.BigDecimal
 import java.time.Instant
 
 data class TransactionDetails(
     val type: String,
     val date: Instant,
-    val valuta: Double,
+    val valuta: BigDecimal,
     val details: String,
 )
