+++ src/main/kotlin/de/babsek/demo/axontesting/projection/BankAccountProjectionEntity.kt
@@ -2,9 +2,14 @@ package de.babsek.demo.axontesting.projection
 
 import com.vladmihalcea.hibernate.type.json.JsonBinaryType
 import de.babsek.demo.axontesting.domain.value.TransactionDetails
-import jakarta.persistence.*
+import jakarta.persistence.Column
+import jakarta.persistence.Entity
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.GenerationType
+import jakarta.persistence.Id
+import jakarta.persistence.Table
 import org.hibernate.annotations.Type
-
+import java.math.BigDecimal
 
 @Table(name = "bank_account_projection")
 @Entity
@@ -21,7 +26,7 @@ data class BankAccountProjectionEntity(
     val ownerName: String,
 
     @Column(name = "balance")
-    val balance: Double,
+    val balance: BigDecimal,
 
     @Type(JsonBinaryType::class)
     @Column(name = "transactions", columnDefinition = "jsonb")
