+++ src/main/kotlin/de/babsek/demo/axontesting/remote/BankAccountDto.kt
@@ -1,10 +1,11 @@
 package de.babsek.demo.axontesting.remote
 
 import de.babsek.demo.axontesting.domain.value.TransactionDetails
+import java.math.BigDecimal
 
 data class BankAccountDto(
     val bankAccountId: String,
     val ownerName: String,
-    val balance: Double,
+    val balance: BigDecimal,
     val transactions: List<TransactionDetails>,
 )
